KanbanBoard/requirements.txt
flask
flask-sqlalchemy
pyodbc
flask-restx
python-dotenv
flask-jwt-extended
werkzeug
python-dateutil